requirements.txt
Click==7.0
Flask-SQLAlchemy==2.4.0
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2==2.8.3
SQLAlchemy==1.3.8
Werkzeug==0.15.5
Pillow==6.1.0